AdBot-Telegram/requirements.txt
python-telegram-bot==21.5
google-generativeai==0.7.2
python-dotenv==1.0.1
requests==2.32.3
Pillow==10.4.0
